use size_t for 64-bit platforms
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@512 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ include/qore/ReferenceHolder.h
@@ -41,13 +41,16 @@
 // ExceptionSink xsink;
 // ReferenceHolder<TibcoClass> holder(self->getReferencePrivateData(...), &xsink);
 //
+
+#include <stdlib.h>
+
 template<typename T>
 class ReferenceHolder
 {
 private:
   ReferenceHolder(const ReferenceHolder&); // not implemented
   ReferenceHolder& operator=(const ReferenceHolder&); // not implemented
-  void* operator new(unsigned); // not implemented, make sure it is not new'ed
+  void* operator new(size_t); // not implemented, make sure it is not new'ed
 
   T* p;
   ExceptionSink* xsink;
